enhance implementation
