+++ src/threedsecure/cardinal/index.ts
@@ -1,1 +1,2 @@
-export * from './cardinal';
+export { init } from './init';
+export { validatePayment } from './validate-payment';
